# api-gateway/openapi-spec.yaml
swagger: "2.0"
info:
  title: Hotel Booking API
  version: "1.0.0"
host: hotel-booking-gateway-8n6yll8a.ew.gateway.dev
schemes:
  - https
produces:
  - application/json
consumes:
  - application/json
...
securityDefinitions:
  firebase:
    authorizationUrl: ""
    flow: implicit
    type: oauth2
    x-google-issuer: "https://securetoken.google.com/hotel-booking-system-final"
    x-google-jwks_uri: "https://www.googleapis.com/service_accounts/v1/jwk/securetoken@system.gserviceaccount.com"
    x-google-audiences: "hotel-booking-system-final"
paths:
  /api/v1/hotels:
    options:
      summary: CORS preflight
      operationId: corsHotels
      x-google-backend:
        address: https://admin-service-677400217786.europe-west3.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      responses:
        200:
          description: OK
    get:
      summary: List hotels
      operationId: listHotels
      x-google-backend:
        address: https://admin-service-677400217786.europe-west3.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      responses:
        200:
          description: OK
    post:
      summary: Create hotel
      operationId: createHotel
      security:
        - firebase: []
      x-google-backend:
        address: https://admin-service-677400217786.europe-west3.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      responses:
        201:
          description: Created
  /api/v1/hotels/{id}:
    options:
      summary: CORS preflight
      operationId: corsHotelById
      x-google-backend:
        address: https://admin-service-677400217786.europe-west3.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        200:
          description: OK
    get:
      summary: Get hotel
      operationId: getHotel
      x-google-backend:
        address: https://admin-service-677400217786.europe-west3.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        200:
          description: OK
    put:
      summary: Update hotel
      operationId: updateHotel
      security:
        - firebase: []
      x-google-backend:
        address: https://admin-service-677400217786.europe-west3.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        200:
          description: OK
    delete:
      summary: Delete hotel
      operationId: deleteHotel
      security:
        - firebase: []
      x-google-backend:
        address: https://admin-service-677400217786.europe-west3.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        200:
          description: OK
...
  /api/v1/search/hotels/{id}:
    options:
      summary: CORS preflight
      operationId: corsSearchHotelById
      x-google-backend:
        address: https://client-service-677400217786.europe-west3.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        200:
          description: OK
    get:
      summary: Get hotel details
      operationId: getHotelDetails
      x-google-backend:
        address: https://client-service-677400217786.europe-west3.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        200:
          description: OK
...